add gymClient

+++ src/main/scala/gym/scala/client/GymSpace.scala
@@ -56,7 +56,7 @@ case class GymObsInfo(info:ObsInfo)
 object GymObsInfo extends DefaultJsonProtocol with SprayJsonSupport {
   implicit val gymInfoFormat = jsonFormat1(GymObsInfo.apply)
 }
-case class GymStepInfo(observation:List[Double], reward:Int, done:Boolean, info:Map[String, String])
+case class GymStepInfo(observation:List[Double], reward:Float, done:Boolean, info:Map[String, String])
 object GymStepInfo extends DefaultJsonProtocol with SprayJsonSupport {
   implicit val gymStepFormat = jsonFormat4(GymStepInfo.apply)
 }
